+++ src/plone/restapi/services/configure.zcml
@@ -1,0 +1,6 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope">
+
+  <include package=".content"/>
+
+</configure>
