+++ src/plone/restapi/services/content/add.py
@@ -7,7 +7,6 @@ from plone.rest import Service
 from plone.restapi.deserializer import DeserializationError
 from plone.restapi.deserializer import json_body
 from plone.restapi.interfaces import IDeserializeFromJson
-from plone.restapi.interfaces import ISerializeToJson
 from random import randint
 from zExceptions import BadRequest
 from zope.component import queryMultiAdapter
@@ -16,42 +15,8 @@ from zope.container.interfaces import INameChooser
 import transaction
 
 
-class DexterityGet(Service):
-
-    def render(self):
-        return ISerializeToJson(self.context)
-
-
-class ContentPatch(Service):
-
-    def render(self):
-        sm = getSecurityManager()
-        if not sm.checkPermission('Modify portal content', self.context):
-            raise Unauthorized
-
-        deserializer = queryMultiAdapter((self.context, self.request),
-                                         IDeserializeFromJson)
-        if deserializer is None:
-            self.request.response.setStatus(501)
-            return dict(error=dict(
-                message='Cannot deserialize type {}'.format(
-                    self.context.portal_type)))
-
-        try:
-            deserializer()
-        except DeserializationError as e:
-            self.request.response.setStatus(400)
-            return dict(error=dict(
-                type='DeserializationError',
-                message=str(e)))
-
-        # TODO: alternativley return the patched object with a 200
-        self.request.response.setStatus(204)
-        return None
-
-
 class FolderPost(Service):
-    """Creates a new object.
+    """Creates a new content object.
     """
 
     def render(self):
@@ -123,32 +88,3 @@ class FolderPost(Service):
         self.request.response.setStatus(201)
         self.request.response.setHeader('Location', obj.absolute_url())
         return None
-
-# class DexterityPost(Service):
-#
-#     def render(self):
-#         return {'service': 'post'}
-
-
-# class DexterityPut(Service):
-#
-#     def render(self):
-#         return {'service': 'put'}
-
-
-# class DexterityDelete(Service):
-#
-#     def render(self):
-#         return {'service': 'delete'}
-
-
-class PloneSiteRootGet(Service):
-
-    def render(self):
-        return ISerializeToJson(self.context)
-
-
-# class PloneSiteRootPost(Service):
-#
-#     def render(self):
-#         return {'service': 'options'}
